Add prompt loader that coordinates commands
The zendesk_search serves as an interactor for all the commands in the
prompt. zendesk_search organizes and coordinates all commands to ensure
a proper flow of the application. Commands are only called from
zendesk_search and a user selected value is returned when a prompt is
asking for input