Create builder to select value converter class
Improve on the way converter type classes are called by using a builder
that determines which type class will convert the search value. Removed
the string type converter as it is the default converter that does not
require any checking.
This pattern enables extension for more converter types without
modifying a long conditional statement.